Moved multiple components to their respective folders and updated the import paths as well; updated dependencies
