+++ app/components/recipe-detail/CollapsibleSection.js
