+++ app/components/recipes.js
@@ -11,7 +11,7 @@ import {
   FaShoppingBag,
 } from "react-icons/fa";
 import { PiCookingPotDuotone, PiHeart } from "react-icons/pi";
-import Carousel from "./Carousel";
+import Carousel from "./ui/Carousel";
 import { SortControl } from "./SortControl";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useShoppingList } from '../context/ShoppingListContext';
